refactor(TestFixture): 패키지 이동
