+++ src/test/java/subway/domain/Line2Test.java
@@ -2,10 +2,10 @@ package subway.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static subway.integration.TestFixture.잠실나루역;
-import static subway.integration.TestFixture.잠실새내역;
-import static subway.integration.TestFixture.잠실역;
-import static subway.integration.TestFixture.종합운동장역;
+import static subway.TestFixture.잠실나루역;
+import static subway.TestFixture.잠실새내역;
+import static subway.TestFixture.잠실역;
+import static subway.TestFixture.종합운동장역;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
